5.0.0/dbHelper (#153)
* [NEW] move dbHelper.js into db folder

* [NEW] add dbHelper alavirien dedicated file

* [UPDATE] index.js

* [NEW] twitter dbHelper file

* [NEW] gift in dbHelper file

* [NEW] ignore in dbHelper file

* [NEW] admin in dbHelper file

* [NEW] birthday in dbHelper file

* [NEW] apology as dbHelper file

* [UPDATE] utils with comments

* [CHANGELOG]

* [PRETTIER]

* [UPDATE] isApologyUser with db as first arg

* [UPDATE] addApologyCount with db as first arg

* [UPDATE] removeApologyCount with db as first arg

* [UPDATE] isBirthdayDate arg order + remove from export

* [UPDATE] birthday.js

* [UPDATE] removeBirthday

* [FIX] missing reference when updating dbAlavirien

* [FIX] missing reference when updating dbApology

* [FIX] missing reference when updating dbBirthday

* [UPDATE] addIgnoredUser with db as first arg

* [UPDATE] removeIgnoredUser with db as first arg

* [UPDATE] isIgnoredUser with db as first arg

* [UPDATE] isTwitterUser with db as first arg

* [UPDATE] getTwitterUser with db as first arg

* [UPDATE] updateLastTweetId with db as first arg

* [UPDATE] addMissingTweets with db as first arg

* [UPDATE] removeMissingTweets with db as first arg

* [UPDATE] remove typo in removeApologyCount

* [DISABLE] removeApologyCount

* [DISABLE] leadApo

* [PRETTIER]

* [CHANGELOG]
