+++ src/commands/slash.js
@@ -113,11 +113,11 @@ const ignoreUser = {
     const iPerso = PERSONALITY.getCommands().ignoreUser;
 
     //check for command argument
-    if (isIgnoredUser(authorId, db)) {
-      removeIgnoredUser(authorId, db);
+    if (isIgnoredUser(db, authorId)) {
+      removeIgnoredUser(db, authorId);
       interactionReply(interaction, iPerso.notIgnored);
     } else {
-      addIgnoredUser(authorId, db);
+      addIgnoredUser(db, authorId);
       interactionReply(interaction, iPerso.ignored);
     }
   },
