+++ src/helpers/db/dbAdmin.js
@@ -1,0 +1,25 @@
+const addAdminLogs = (db, messageId, type, index) => {
+  const adminLogs = db.data.adminLogs;
+  //{frequent: [[]...], userAD: [[]...]}
+  const data = adminLogs[type]; // [[]...]
+
+  data[index].push(messageId);
+  db.wasUpdated = true;
+};
+
+const getAdminLogs = (db) => {
+  return db.data.adminLogs;
+};
+
+const removeAdminLogs = (db, type) => {
+  const adminLogs = db.data.adminLogs;
+  //{frequent: [[]...], userAD: [[]...]}
+  const data = adminLogs[type]; // [[]...]
+  const sliced = data.slice(1); //remove first
+  sliced.push([]); //add [] at the end
+
+  db.data.adminLogs[type] = sliced;
+  db.wasUpdated = true;
+};
+
+export { addAdminLogs, getAdminLogs, removeAdminLogs };
