+++ src/helpers/db/dbApology.js
@@ -1,0 +1,40 @@
+const isApologyUser = (db, authorId) => {
+  return db.data.apologiesCounting
+    .map((obj) => {
+      return obj.userId;
+    })
+    .includes(authorId);
+};
+
+const addApologyCount = (db, authorId) => {
+  const { apologiesCounting } = db.data;
+
+  if (isApologyUser(db, authorId)) {
+    // If already in DB, add +1 to the counter
+    for (const obj of apologiesCounting) {
+      if (obj.userId === authorId) {
+        obj.counter++;
+      }
+    }
+  } else {
+    // Else add user
+    db.data.apologiesCounting = [
+      ...db.data.apologiesCounting,
+      { userId: authorId, counter: 1 },
+    ];
+  }
+  db.wasUpdated = true;
+};
+
+/*
+const removeApologyCount = (db, authorId) => {
+  if (isApologyUser(db, authorId)) {
+    db.data.apologiesCounting = db.data.apologiesCounting.filter(
+      (cur) => cur.userId !== authorId
+    );
+    db.wasUpdated = true;
+  }
+};
+*/
+
+export { isApologyUser, addApologyCount }; //, removeApologyCount };
