+++ src/helpers/db/dbIgnore.js
@@ -1,0 +1,45 @@
+//IGNORE CHANNEL
+const isIgnoredChannel = (db, channelId) => {
+  return db.data.ignoredChannelIds.includes(channelId);
+};
+
+const addIgnoredChannel = (db, channelId) => {
+  if (!isIgnoredChannel(db, channelId)) {
+    db.data.ignoredChannelIds.push(channelId);
+  }
+  db.wasUpdated = true;
+};
+
+const removeIgnoredChannel = (db, channelId) => {
+  if (isIgnoredChannel(db, channelId)) {
+    db.data.ignoredChannelIds = db.data.ignoredChannelIds.filter(
+      (id) => id !== channelId
+    );
+  }
+  db.wasUpdated = true;
+};
+
+export { isIgnoredChannel, addIgnoredChannel, removeIgnoredChannel };
+
+// IGNORED USERS
+const addIgnoredUser = (db, authorId) => {
+  if (!db.data.ignoredUsersIds.includes(authorId)) {
+    db.data.ignoredUsersIds.push(authorId);
+  }
+  db.wasUpdated = true;
+};
+
+const removeIgnoredUser = (db, authorId) => {
+  if (db.data.ignoredUsersIds.includes(authorId)) {
+    db.data.ignoredUsersIds = db.data.ignoredUsersIds.filter(
+      (id) => id !== authorId
+    );
+  }
+  db.wasUpdated = true;
+};
+
+const isIgnoredUser = (db, authorId) => {
+  return db.data.ignoredUsersIds.includes(authorId);
+};
+
+export { addIgnoredUser, removeIgnoredUser, isIgnoredUser };
